+++ bin/metrics.dart
@@ -1,90 +1,44 @@
 import 'dart:io';
 
-import 'package:args/args.dart';
 import 'package:dart_code_metrics/metrics_analyzer.dart';
 import 'package:dart_code_metrics/reporters.dart';
+import 'package:dart_code_metrics/src/cli/arguments_parser.dart';
+import 'package:dart_code_metrics/src/cli/arguments_validation.dart';
+import 'package:dart_code_metrics/src/cli/arguments_validation_exceptions.dart';
 import 'package:glob/glob.dart';
 
-const _usageHeader = 'Usage: metrics [options...] <directories>';
-const _helpFlagName = 'help';
-const _reporterOptionName = 'reporter';
-const _cyclomaticComplexityThreshold = 'cyclomatic-complexity';
-const _linesOfCodeThreshold = 'lines-of-code';
-const _verboseName = 'verbose';
-const _ignoredFilesName = 'ignore-files';
-const _rootFolderName = 'root-folder';
-
-final _parser = ArgParser()
-  ..addFlag(_helpFlagName,
-      abbr: 'h', help: 'Print this usage information.', negatable: false)
-  ..addOption(_reporterOptionName,
-      abbr: 'r',
-      help: 'The format of the output of the analysis',
-      valueHelp: 'console',
-      allowed: ['console', 'json', 'html', 'codeclimate'],
-      defaultsTo: 'console')
-  ..addOption(_cyclomaticComplexityThreshold,
-      help: 'Cyclomatic complexity threshold',
-      valueHelp: '20',
-      defaultsTo: '20', callback: (String i) {
-    if (int.tryParse(i) == null) print('$_cyclomaticComplexityThreshold:');
-  })
-  ..addOption(_linesOfCodeThreshold,
-      help: 'Lines of code threshold',
-      valueHelp: '50',
-      defaultsTo: '50', callback: (String i) {
-    if (int.tryParse(i) == null) print('$_linesOfCodeThreshold:');
-  })
-  ..addOption(_rootFolderName,
-      help: 'Root folder', valueHelp: './', defaultsTo: Directory.current.path)
-  ..addOption(_ignoredFilesName,
-      help: 'Filepaths in Glob syntax to be ignored',
-      valueHelp: '{/**.g.dart,/**.template.dart}',
-      defaultsTo: '{/**.g.dart,/**.template.dart}')
-  ..addFlag(_verboseName, negatable: false);
+final parser = argumentsParser();
 
 void main(List<String> args) {
   try {
-    final arguments = _parser.parse(args);
+    final arguments = parser.parse(args);
 
-    if (arguments[_helpFlagName] as bool) {
+    if (arguments[helpFlagName] as bool) {
       _showUsageAndExit(0);
     }
 
-    if (arguments.rest.isEmpty) {
-      throw _InvalidArgumentException(
-          'Invalid number of directories. At least one must be specified');
-    }
-
-    arguments.rest.forEach((p) {
-      if (!Directory(p).existsSync()) {
-        throw _InvalidArgumentException(
-            "$p doesn't exist or isn't a directory");
-      }
-    });
-
-    // TODO: check that directories to analyze are all children of root folder
+    validateArguments(arguments);
 
     _runAnalysis(
-        arguments[_rootFolderName] as String,
+        arguments[rootFolderName] as String,
         arguments.rest,
-        arguments[_ignoredFilesName] as String,
-        int.parse(arguments[_cyclomaticComplexityThreshold] as String),
-        int.parse(arguments[_linesOfCodeThreshold] as String),
-        arguments[_reporterOptionName] as String,
-        arguments[_verboseName] as bool);
+        arguments[ignoredFilesName] as String,
+        int.parse(arguments[cyclomaticComplexityThreshold] as String),
+        int.parse(arguments[linesOfCodeThreshold] as String),
+        arguments[reporterOptionName] as String,
+        arguments[verboseName] as bool);
   } on FormatException catch (e) {
     print('${e.message}\n');
     _showUsageAndExit(1);
-  } on _InvalidArgumentException catch (e) {
+  } on InvalidArgumentException catch (e) {
     print('${e.message}\n');
     _showUsageAndExit(1);
   }
 }
 
 void _showUsageAndExit(int exitCode) {
-  print(_usageHeader);
-  print(_parser.usage);
+  print(usageHeader);
+  print(parser.usage);
   exit(exitCode);
 }
 
@@ -138,9 +92,4 @@ void _runAnalysis(
   }
 
   reporter.report(runner.results()).forEach(print);
-}
-
-class _InvalidArgumentException implements Exception {
-  final String message;
-  _InvalidArgumentException(this.message);
 }