+++ lib/src/cli/arguments_validation_exceptions.dart
@@ -1,0 +1,4 @@
+class InvalidArgumentException implements Exception {
+  final String message;
+  InvalidArgumentException(this.message);
+}