+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+# 1.2.1
+- Validate root-folder argument
+- Fix paths to analyze fail to validate with non-default root-folder
+- Fix paths weren't validated to be inside root-folder
+
 # 1.2.0
 - Allow to analyze multiple directories